Teacher able to delete Assignment except assignments already assigned

+++ app/controllers/assignments_controller.rb
@@ -28,10 +28,14 @@ end
 
 def destroy
   assignment =Assignment.find_by(id: params[:id])
-
   if assignment 
-    assignment.destroy
-    head :no_content 
+    student_assignments = StudentAssignment.where(assignment_id: assignment.id)
+    if student_assignments.length > 0
+      render json: {errors: ["Cannot Delete Assignment right now as there are students already assigned to it"]}
+    else
+      assignment.destroy
+      render json: assignment
+    end
   else
    render json: {error: "Assignment not found"}, status: :not_found
   end
